requirements/base.txt
pytz==2018.5
redis>=2.10.5
celery==4.2.1
flower==0.9.2
psycopg2==2.7.4

# Django
# ------------------------------------------------------------------------------
django==1.11.15
django-environ==0.4.5
django-filter==2.0.0
django-model-utils==3.1.2
django-redis==4.9.0
django-phonenumber-field==2.0.1

# Django REST Framework
coreapi==2.3.3
djangorestframework==3.8.2
